problem: 202
Upon sketching out this problem, it seemed similiar to finding cycle in a linked list. So I reminded myself the Floyd's cycle finding algorithm (or Hare-Tortoise algorithm) which I've already implemented for problem in the past.

It turned out the solution was correct and also it was one of the most intuitive and clever solutions based on the submissions/comments of others.

I was just so happy that I've found the solution so quickly. It was a really good pattern recognition and analogical thinking!