problem: 203
I spent way more time on this problem than I should've. The most elegant and quick solution is using recursion, but I haven't recognized this and opted for the solution using some pointer magic.

Actually it took me long time to understand after I've come up with this solution. But we basically need to keep double pointer (pointer to pointer) to be able to bypass any current node we are on for the next one and also handle edge cases when head needs to be bypassed or when multiple consecutive elements need to be bypassed. And we need pointer to pointer of node because otherwise the pointer to node is copied by value and doesn't change the original, previous node's->next pointer.

Other than the complexity, I have to give myself credit - this solution is fast and quite clever, I haven't found anybody else using this solution.